rename NavBar to HomeNavBar
