+++ client/src/components/styles/Login.styled.ts
@@ -5,7 +5,7 @@ import StyledButton from './Button.styled';
 import StyledInput from './TextField.styled';
 
 export const StyledLogin = styled.div`
-height: 100vh;
+min-height: 100vh;
 display: flex;
 justify-content: center;
 align-items: center;
